Add common RunSummaryFormatter base class
Former RunSummaryFormatter has been renamed to
ImportantRunSummaryFormatter.
